Move compose to common
